+++ src/lib/components/Accordion/AccordionPanel.svelte
@@ -5,8 +5,6 @@
   export let panelID;
   export let labeledBy;
   export let style;
-
-  console.log(style);
 </script>
 
 <div
